Rest Framework Tutorial/backend/products/views.py
from rest_framework import generics

from .models import Product
from .serializers import ProductSerializer

# For Function Based View
from rest_framework.response import Response
from rest_framework.decorators import api_view
from django.shortcuts import get_object_or_404


class ProductListCreateAPIView(generics.ListCreateAPIView):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer

    def perform_create(self, serializer):
        content = serializer.validated_data.get('content')
        if content is None:
            content = "created using a generic view"
        serializer.save(content=content)


product_list_create_view = ProductListCreateAPIView.as_view()


class ProductDetailAPIView(generics.RetrieveAPIView):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer


product_detail_view = ProductDetailAPIView.as_view()


class ProductUpdateAPIView(generics.UpdateAPIView):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer
    lookup_field = 'pk'


product_update_view = ProductUpdateAPIView.as_view()


class ProductDeleteAPIView(generics.DestroyAPIView):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer
    lookup_field = 'pk'


product_delete_view = ProductDeleteAPIView.as_view()


# Alternative Function Based View
@api_view(["GET", "POST"])
def all_function_view(request, pk=None, *args, **kwargs):
    if request.method == "GET":
        if pk is None:
            queryset = Product.objects.all()
            serializer = ProductSerializer(queryset, many=True)
            return Response(serializer.data)
        obj = get_object_or_404(Product, pk=pk)
        serializer = ProductSerializer(obj, many=False)
        return Response(serializer.data)
    if request.method == "POST":
        serializer = ProductSerializer(data=request.data)
        if serializer.is_valid(raise_exception=True):
            content = serializer.validated_data.get('content')
            if content is None:
                content = "created using a functional view"
            serializer.save(content=content)
        return Response(serializer.data)
