Rest Framework Tutorial/requirements.txt
pyyaml
requests
django
djangorestframework
django-cors-headers